Fix GPG_TTY export command for non-bash shells
I fell foul of an issue with zsh power10k theme as described in this article. https://unix.stackexchange.com/questions/608842/zshrc-export-gpg-tty-tty-says-not-a-tty. 

This result was that a first test of signing a commit resulted in terminal output saying the commit signing failed without ever having prompted for a passphrase.

The change suggested worked for me and is according to the article should work for many people.